+++ src/components/KFileUpload/KFileUpload.vue
@@ -293,8 +293,9 @@ export default defineComponent({
     height: 29px;
   }
 
-  // To hide the thumbnail that appears in Safari after uploading a file
-  :deep(.k-input-wrapper) input[type="file"]::-webkit-file-upload-button {
+  // To hide the button and thumbnail that appears in Safari and firefox after uploading a file
+  :deep(.k-input-wrapper) input[type="file"]::-webkit-file-upload-button,
+  :deep(.k-input-wrapper) input[type="file"]::file-selector-button {
     position: absolute;
     min-width: 100%;
     min-height: 100%;